Perform refactoring
