+++ javaTodoApp/src/model/Task.java
@@ -1,4 +1,4 @@
-package main;
+package model;
 
 /*
 * Userが登録するTaskクラス
